- getting rid of more migration leftovers - scaffolding SchemaEncoder.Encode to refactor out IDTH.CreateThrift()
